Remove old HomeActivity and its resources
