--- app/src/main/res/drawable/vector_disable_able_gradient_button_bg_external_dark.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="oval">
-            <solid android:color="#80000000" />
-            <stroke android:color="#80613B62" android:width="1dp"/>
-            <corners android:radius="39dp" />
-        </shape>
-    </item>
-</layer-list>
