--- app/src/main/res/drawable/vector_disable_able_gradient_button_bg_round.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_enabled="true">
-        <shape android:shape="rectangle">
-            <solid android:color="?attr/palette_button_primary" />
-            <corners android:radius="30dp" />
-        </shape>
-    </item>
-    <item android:state_enabled="false">
-        <shape android:shape="rectangle">
-            <solid android:color="?attr/palette_button_disable" />
-            <corners android:radius="30dp" />
-        </shape>
-    </item>
-</selector>
