--- app/src/main/res/drawable/vector_home_book.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M2.25,3.75C5.29,3.764 7.408,4.146 8.833,4.827C10.183,5.472 10.92,6.387 11.25,7.597V19.237C9.545,17.978 6.923,17.251 2.264,17.25C2.258,17.233 2.25,17.204 2.25,17.159V3.75ZM12.75,19.236C14.459,17.972 17.079,17.25 21.75,17.25V3.75C18.711,3.764 16.592,4.146 15.167,4.827C13.817,5.472 13.08,6.387 12.75,7.597V19.236ZM12,5.611C11.431,4.723 10.599,4.008 9.48,3.473C7.76,2.651 5.379,2.264 2.253,2.25L2.253,2.25C2.055,2.249 1.86,2.288 1.677,2.363C1.494,2.438 1.328,2.549 1.189,2.688C1.049,2.828 0.938,2.994 0.863,3.177C0.788,3.359 0.749,3.554 0.75,3.751V17.159C0.75,17.541 0.868,17.932 1.13,18.24C1.402,18.558 1.798,18.75 2.25,18.75C8.161,18.75 10.201,19.951 11.414,21.468C11.556,21.646 11.772,21.75 12,21.75C12.228,21.75 12.443,21.646 12.586,21.468C13.807,19.942 15.839,18.75 21.75,18.75C22.148,18.75 22.529,18.592 22.811,18.31C23.092,18.029 23.25,17.648 23.25,17.25V3.751C23.25,3.554 23.212,3.359 23.137,3.177C23.062,2.994 22.951,2.828 22.812,2.688C22.672,2.549 22.506,2.438 22.323,2.363C22.14,2.288 21.945,2.249 21.747,2.25L21.747,2.25C18.621,2.264 16.24,2.651 14.52,3.473C13.401,4.008 12.569,4.723 12,5.611Z"
-      android:fillColor="#0C0C0E"
-      android:fillType="evenOdd"/>
-</vector>
