--- app/src/main/res/drawable/vector_home_chat.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M4.873,2.25L4.875,2.25H19.125V3L19.127,2.25C20.021,2.253 20.877,2.609 21.509,3.241C22.141,3.873 22.497,4.729 22.5,5.623L22.5,5.625V14.625L22.5,14.627C22.497,15.521 22.141,16.377 21.509,17.009C20.877,17.641 20.021,17.997 19.127,18L19.125,18H11.519L7.23,21.576C7.007,21.762 6.696,21.803 6.432,21.679C6.168,21.556 6,21.291 6,21V18H4.875L4.873,18C3.979,17.997 3.123,17.641 2.491,17.009C1.859,16.377 1.503,15.521 1.5,14.627L1.5,14.625V5.625L1.5,5.623C1.503,4.729 1.859,3.873 2.491,3.241C3.123,2.609 3.979,2.253 4.873,2.25ZM11.624,17.913C11.624,17.913 11.624,17.913 11.623,17.913L11.624,17.913ZM19.124,3.75H4.876C4.379,3.752 3.903,3.95 3.551,4.301C3.2,4.653 3.002,5.129 3,5.626V14.624C3.002,15.121 3.2,15.597 3.551,15.949C3.903,16.3 4.379,16.498 4.876,16.5H6.75C7.164,16.5 7.5,16.836 7.5,17.25V19.398L10.662,16.762C10.663,16.761 10.663,16.761 10.663,16.761C10.865,16.592 11.12,16.5 11.384,16.5C11.384,16.5 11.384,16.5 11.384,16.5H19.124C19.621,16.498 20.097,16.3 20.449,15.949C20.8,15.597 20.998,15.121 21,14.624V5.626C20.998,5.129 20.8,4.653 20.449,4.301C20.097,3.95 19.621,3.752 19.124,3.75Z"
-      android:fillColor="#0C0C0E"
-      android:fillType="evenOdd"/>
-</vector>
