--- app/src/main/res/drawable/vector_home_home.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M12.261,4.747C12.191,4.681 12.098,4.643 12.001,4.643C11.905,4.643 11.812,4.681 11.742,4.747L3.113,12.991C3.076,13.026 3.047,13.069 3.027,13.115C3.007,13.162 2.997,13.212 2.997,13.263L2.995,21C2.995,21.398 3.153,21.779 3.435,22.061C3.716,22.342 4.098,22.5 4.495,22.5H9C9.199,22.5 9.39,22.421 9.53,22.28C9.671,22.14 9.75,21.949 9.75,21.75V15.375C9.75,15.276 9.79,15.18 9.86,15.11C9.93,15.04 10.026,15 10.125,15H13.875C13.974,15 14.07,15.04 14.14,15.11C14.21,15.18 14.25,15.276 14.25,15.375V21.75C14.25,21.949 14.329,22.14 14.47,22.28C14.61,22.421 14.801,22.5 15,22.5H19.503C19.901,22.5 20.282,22.342 20.563,22.061C20.845,21.779 21.003,21.398 21.003,21V13.263C21.003,13.212 20.993,13.162 20.973,13.115C20.953,13.069 20.924,13.026 20.887,12.991L12.261,4.747Z"
-      android:fillColor="#9330FF"/>
-  <path
-      android:pathData="M23.011,11.444L19.505,8.09V3C19.505,2.801 19.426,2.61 19.285,2.47C19.145,2.329 18.954,2.25 18.755,2.25H16.505C16.306,2.25 16.115,2.329 15.975,2.47C15.834,2.61 15.755,2.801 15.755,3V4.5L13.04,1.904C12.786,1.647 12.408,1.5 12,1.5C11.593,1.5 11.216,1.647 10.962,1.905L0.992,11.444C0.7,11.725 0.664,12.188 0.929,12.492C0.996,12.569 1.077,12.632 1.169,12.676C1.26,12.72 1.36,12.746 1.462,12.75C1.563,12.755 1.665,12.739 1.76,12.703C1.855,12.667 1.942,12.612 2.016,12.542L11.742,3.247C11.812,3.181 11.905,3.143 12.001,3.143C12.098,3.143 12.191,3.181 12.261,3.247L21.988,12.542C22.131,12.679 22.323,12.754 22.522,12.75C22.72,12.747 22.909,12.664 23.047,12.521C23.335,12.223 23.311,11.731 23.011,11.444Z"
-      android:fillColor="#9330FF"/>
-</vector>
