--- app/src/main/res/drawable/vector_ttl_store_icon.xml
@@ -1,18 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="25dp"
-    android:height="22dp"
-    android:viewportWidth="25"
-    android:viewportHeight="22">
-  <group>
-    <clip-path
-        android:pathData="M0.928,0.001L24.332,0.001 24.332,15.917 0.928,15.917z"/>
-    <path
-        android:pathData="M22.76,6.185l-1.849,7.59c-0.098,0.37 -0.419,0.615 -0.788,0.615h-9.83c-0.37,0 -0.69,-0.246 -0.79,-0.616l-2.143,-8.6h14.611c0.246,0 0.493,0.124 0.641,0.321 0.147,0.197 0.221,0.444 0.147,0.69m1.06,-1.65c-0.444,-0.568 -1.134,-0.913 -1.848,-0.913L6.965,3.622L6.497,1.8C6.227,0.74 5.29,0 4.207,0L1.692,0C1.274,0 0.928,0.345 0.928,0.764c0,0.42 0.346,0.764 0.765,0.764h2.513c0.37,0 0.69,0.247 0.788,0.616L8,14.12c0.271,1.06 1.207,1.798 2.292,1.798h9.831c1.084,0 2.045,-0.739 2.291,-1.798l1.848,-7.589c0.173,-0.69 0.025,-1.429 -0.443,-1.996"
-        android:fillColor="#000"
-        android:fillType="evenOdd"/>
-  </group>
-  <path
-      android:pathData="M10.81,20.476c-0.567,0 -1.01,-0.443 -1.01,-1.01s0.443,-1.01 1.01,-1.01c0.566,0 1.01,0.443 1.01,1.01 0,0.541 -0.469,1.01 -1.01,1.01m0,-3.573c-1.405,0 -2.564,1.158 -2.564,2.562 0,1.406 1.16,2.563 2.563,2.563 1.405,0 2.563,-1.157 2.563,-2.563 0,-1.404 -1.158,-2.562 -2.563,-2.562M19.285,20.476c-0.566,0 -1.01,-0.443 -1.01,-1.01s0.444,-1.01 1.01,-1.01c0.567,0 1.01,0.443 1.01,1.01 0,0.541 -0.468,1.01 -1.01,1.01m0,-3.573c-1.404,0 -2.563,1.158 -2.563,2.562 0,1.406 1.16,2.563 2.563,2.563 1.404,0 2.563,-1.157 2.563,-2.563 -0.025,-1.404 -1.159,-2.562 -2.563,-2.562"
-      android:fillColor="#000"
-      android:fillType="evenOdd"/>
-</vector>
