--- app/src/main/res/layout/activity_home_old.xml
@@ -1,160 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/home_activity_view_root"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <!--     TODO remove, and remove resources-->
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/home_root"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:animateLayoutChanges="true">
-
-        <androidx.viewpager2.widget.ViewPager2
-            android:id="@+id/view_pager"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            app:layout_constraintBottom_toTopOf="@+id/bottom_navigation_view"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="1.0"
-            tools:ignore="FragmentTagUsage" />
-
-        <View
-            android:id="@+id/bottom_elevation_view"
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/common_scroll_height_gradient_height"
-            android:alpha="0.7"
-            android:background="@drawable/vector_view_elevation_bottom_gradient"
-            app:layout_constraintBottom_toTopOf="@id/bottom_navigation_view" />
-
-        <com.tari.android.wallet.ui.component.tari.background.TariPrimaryBackground
-            android:id="@+id/bottom_navigation_view"
-            android:layout_width="0dp"
-            android:layout_height="64dp"
-            android:elevation="0dp"
-            android:orientation="horizontal"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:orientation="horizontal">
-
-                <FrameLayout
-                    android:id="@+id/home_view"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="1">
-
-                    <com.tari.android.wallet.ui.component.tari.TariIconView
-                        android:id="@+id/home_image_view"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center"
-                        android:src="@drawable/vector_home_home" />
-                </FrameLayout>
-
-                <FrameLayout
-                    android:id="@+id/store_view"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="1">
-
-                    <com.tari.android.wallet.ui.component.tari.TariIconView
-                        android:id="@+id/store_image_view"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center"
-                        android:src="@drawable/vector_home_book" />
-                </FrameLayout>
-
-                <View
-                    android:layout_width="50dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="0.35" />
-
-                <FrameLayout
-                    android:id="@+id/chat_view"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="1">
-
-                    <com.tari.android.wallet.ui.component.tari.TariIconView
-                        android:id="@+id/chat_image_view"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center"
-                        android:src="@drawable/vector_home_chat" />
-                </FrameLayout>
-
-                <FrameLayout
-                    android:id="@+id/settings_view"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="1">
-
-                    <com.tari.android.wallet.ui.component.tari.TariIconView
-                        android:id="@+id/settings_image_view"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center"
-                        android:src="@drawable/vector_home_menu" />
-                </FrameLayout>
-
-            </LinearLayout>
-
-        </com.tari.android.wallet.ui.component.tari.background.TariPrimaryBackground>
-
-        <com.tari.android.wallet.ui.component.tari.background.TariAlphaBackground
-            android:id="@+id/send_button_container"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="-40dp"
-            android:elevation="30dp"
-            app:backgroundColor="?attr/palette_background_primary"
-            app:cornerRadius="40dp"
-            app:elevation="30dp"
-            app:layout_constraintBottom_toTopOf="@id/bottom_navigation_view"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent">
-
-            <androidx.constraintlayout.widget.ConstraintLayout
-                android:id="@+id/send_button_external_container"
-                android:layout_width="78dp"
-                android:layout_height="78dp"
-                android:background="@drawable/vector_disable_able_gradient_button_bg_external">
-
-                <androidx.appcompat.widget.AppCompatImageView
-                    android:id="@+id/send_tari_button"
-                    android:layout_width="60dp"
-                    android:layout_height="60dp"
-                    android:layout_marginHorizontal="9dp"
-                    android:layout_marginVertical="9dp"
-                    android:background="@drawable/vector_disable_able_gradient_button_bg_round"
-                    android:scaleType="center"
-                    android:src="@drawable/vector_icon_send_tari"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-            </androidx.constraintlayout.widget.ConstraintLayout>
-
-        </com.tari.android.wallet.ui.component.tari.background.TariAlphaBackground>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <FrameLayout
-        android:id="@+id/nav_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-</FrameLayout>
